:truck: Move: 공통 컴포넌트 폴더 구조 변경
Move: 공통 컴포넌트 폴더 구조 변경
