+++ app/_components/Button/WishButton.tsx
@@ -11,7 +11,7 @@ import { handleLikeProject } from "@/app/_apis/handleLikeProject";
 import { userQueryKeys } from "@/app/_queryFactory/userQuery";
 import useModal from "@/app/_hooks/useModal";
 import useCheckLogin from "@/app/_hooks/useCheckLogin";
-import LoginModal from "../LoginModal/LoginModal";
+import LoginModal from "../Modal/LoginModal";
 
 interface WishButtonAndCountProps {
   isFavorite: boolean;
