+++ app/_components/LandingPage/AnimationSection.tsx
