+++ app/_components/ProjectList/ProjectCard/HoverCard.tsx
@@ -1,5 +1,5 @@
 import { ProjectData } from "@/app/_apis/schema/projectResponse";
-import WishButtonAndCount from "../../WishButtonAndCount/WishButtonAndCount";
+import WishButtonAndCount from "../../Button/WishButton";
 import StackBadge from "./StackBadge";
 
 function HoverCard({ project }: { project: ProjectData }) {
